Improvements and bug fixes.

+++ tests/my_tests.c
@@ -23,7 +23,7 @@ void put_value(uint64_t n, uint64_t v) {
 }
 
 // Tę funkcję woła rdzeń.
-void print_register(uint64_t n, char v, uint64_t u) {
+void print_register(uint64_t n, char v, long long u) {
     printf("Core %lu printed value %c and %lu\n", n, v, u);
 }
